make playlist additions work for multi users
git-svn-id: https://vdr-muggle.svn.sourceforge.net/svnroot/vdr-muggle/branches/0.1.3-wr@510 e10066b5-e1e2-0310-b819-94efdf66514b

+++ mg_order.h
@@ -132,7 +132,7 @@ sql_string (MYSQL *db, const string s);
 
 MYSQL_RES * exec_sql (MYSQL *db,string query);
 string get_col0 (MYSQL *db,string query);
-int exec_count (MYSQL *db,string query);
+long exec_count (MYSQL *db,string query);
 
 //! \brief converts long to string
 string itos (int i);
